+++ packages/react/src/Tabs/Tab/Tab.types.ts
@@ -1,4 +1,5 @@
 import * as React from "react";
+
 import { TabProps as MaterialTabProps } from "@material-ui/core/Tab";
 
 type OmittedProps = |