+++ packages/flutter/CHANGELOG.md
@@ -3,6 +3,37 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.4.0](https://github.com/FacioCode/design/compare/v4.3.2...v4.4.0) (2021-02-03)
+
+**Note:** Version bump only for package facio_design_system
+
+
+
+
+
+## [4.3.2](https://github.com/FacioCode/design/compare/v4.3.1...v4.3.2) (2021-02-03)
+
+
+### Bug Fixes
+
+* **flutter:** add missing disabled state style for small buttons ([d3d518f](https://github.com/FacioCode/design/commit/d3d518f9bad0520396bdd5e3fb3bcea46e876b5e))
+
+
+
+
+
+## [4.3.1](https://github.com/FacioCode/design/compare/v4.3.0...v4.3.1) (2021-02-02)
+
+
+### Bug Fixes
+
+* **flutter:** do not allow onPress if button is disabled ([45385dd](https://github.com/FacioCode/design/commit/45385ddb2843aaca94a483fa1d36957b3e2dbc71))
+* **flutter:** proper keeping and cleaning states for input text with masks ([8bb5d75](https://github.com/FacioCode/design/commit/8bb5d75441b35e88fa703208b0e27456e7ae585d))
+
+
+
+
+
 # [4.3.0](https://github.com/FacioCode/design/compare/v4.2.0...v4.3.0) (2021-01-30)