+++ packages/style-dictionary/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.4.0](https://github.com/FacioCode/design/compare/v4.3.2...v4.4.0) (2021-02-03)
+
+**Note:** Version bump only for package @faciocode/style-dictionary
+
+
+
+
+
+## [4.3.2](https://github.com/FacioCode/design/compare/v4.3.1...v4.3.2) (2021-02-03)
+
+**Note:** Version bump only for package @faciocode/style-dictionary
+
+
+
+
+
+## [4.3.1](https://github.com/FacioCode/design/compare/v4.3.0...v4.3.1) (2021-02-02)
+
+**Note:** Version bump only for package @faciocode/style-dictionary
+
+
+
+
+
 # [4.3.0](https://github.com/FacioCode/design/compare/v4.2.0...v4.3.0) (2021-01-30)